+++ src/jyut-dict/logic/entry/entryspeaker.cpp
@@ -1,6 +1,6 @@
 #include "entryspeaker.h"
 #include "logic/utils/cantoneseutils.h"
-#include "logic/utils/chineseutils.h"
+#include "logic/utils/mandarinutils.h"
 
 #include <QCoreApplication>
 #include <QFileInfo>
@@ -210,7 +210,7 @@ int EntrySpeaker::speak(const QLocale::Language &language,
             QString mutableString = string;
             mutableString.replace("u:", "ü");
             mutableString.replace("v", "ü");
-            ChineseUtils::segmentPinyin(mutableString, syllables);
+            MandarinUtils::segmentPinyin(mutableString, syllables);
         }
 
         for (const auto &syllable : syllables) {
