+++ src/jyut-dict/logic/utils/scriptdetector.cpp
@@ -1,7 +1,7 @@
 #include "scriptdetector.h"
 
 #include "logic/utils/cantoneseutils.h"
-#include "logic/utils/chineseutils.h"
+#include "logic/utils/mandarinutils.h"
 
 #include <QList>
 #include <QRegularExpression>
@@ -841,10 +841,10 @@ ScriptDetector::ScriptDetector(const QString &string)
     _isValidJyutpingAfterAutocorrect = CantoneseUtils::segmentJyutping(out,
                                                                      dummyVec);
     QString processedPinyin = string;
-    _isValidPinyin = ChineseUtils::segmentPinyin(processedPinyin
-                                                     .replace("v", "u:")
-                                                     .replace("ü", "u:"),
-                                                 dummyVec);
+    _isValidPinyin = MandarinUtils::segmentPinyin(processedPinyin
+                                                      .replace("v", "u:")
+                                                      .replace("ü", "u:"),
+                                                  dummyVec);
 }
 
 bool ScriptDetector::containsChinese()
